Extract the declaration of the deck to a file
In order to enable different decks, this allows the caller to specify
a file containing definitions for a different list of cards.

Note the current default 'deck.env' file is a link to the deck-ffpc.env
file.  Future releases may elect to change the default deck to point to
a different file.